extract fhirAuth
